Change style/color and fix bug
The bug fixed in this commit has to do with the calculator as expected
when the user tried operate two values, with 0 being the second.
This was being caused because, to verify wether we had two valid values
not false; however, thichecked also coounted 0 as false, since
it's a falsy value. So i changed to explicitly check for null.